Split world data loading from runtime updates
